+++ islands/TimelineEvent.tsx
@@ -5,24 +5,30 @@ import Text from "lunchbox/components/Text/index.tsx";
 import Code from "lunchbox/components/Code/index.tsx";
 
 const ModuleContent = (event: MdTimelineEvent) => (
-  <Module size="sm" half="lg">
-    <div class="mt-8 mb-16">
-      <Text type={event.isPrimary ? "heading" : "subheading"} noMargins>
+  <Module size="md" half="lg">
+    <div class={event.isPrimary ? "my-8" : "my-4"}>
+      <Text type={event.isPrimary ? "heading" : "paragraph"} noMargins>
         {event.title}
       </Text>
       <Text type="small" noMargins>{event.role}</Text>
-      <Text noMargins>{event.summary}</Text>
-      <Text>
-        {Array.isArray(event.tags) ? event.tags.join(", ") : event.tags}
-      </Text>
+      {event.isPrimary
+        ? (
+          <>
+            <Text noMargins>{event.summary}</Text>
+            <Text>
+              {Array.isArray(event.tags) ? event.tags.join(", ") : event.tags}
+            </Text>
+          </>
+        )
+        : null}
     </div>
   </Module>
 );
 
 const ModulePeriod = (event: MdTimelineEvent) => (
-  <Module size="sm" half="xs">
-    <div class={event.isPrimary ? "mt-8" : "mt-12"}>
-      <Text class={event.isPrimary ? "text-left" : "text-right"}>
+  <Module size="xs" half="xs">
+    <div class={event.isPrimary ? "mt-9" : "mt-5"}>
+      <Text class="text-right" noMargins>
         {event.period}
       </Text>
     </div>
@@ -32,13 +38,29 @@ const ModulePeriod = (event: MdTimelineEvent) => (
 const ModuleLine = (event: MdTimelineEvent) => (
   <Module size="xs" half="xs" class="relative">
     <div class="h-full w-1/2 border-r border-r-1" />
-    <div
-      class="absolute top-8 w-12 h-12 rounded-full"
-      style={{
-        backgroundColor: "var(--clr-txt-base)",
-        left: "calc(50% - 1.5rem)",
-      }}
-    />
+    {event.isPrimary
+      ? (
+        <div
+          class="p-2 border absolute top-8 rounded-full"
+          style={{ left: "calc(50% - 1.5rem)" }}
+        >
+          <div
+            class="w-8 h-8 rounded-full"
+            style={{
+              backgroundColor: "var(--clr-txt-base)",
+            }}
+          />
+        </div>
+      )
+      : (
+        <div
+          class="absolute top-6 w-4 h-4 rounded-full"
+          style={{
+            backgroundColor: "var(--clr-txt-base)",
+            left: "calc(50% - 0.5rem)",
+          }}
+        />
+      )}
   </Module>
 );
 
@@ -47,13 +69,10 @@ export default function (
 ) {
   return (
     <>
-      {event.attrs.isPrimary
-        ? <ModuleContent {...event.attrs} />
-        : <ModulePeriod {...event.attrs} />}
+      <ModulePeriod {...event.attrs} />
       <ModuleLine {...event.attrs} />
-      {event.attrs.isPrimary
-        ? <ModulePeriod {...event.attrs} />
-        : <ModuleContent {...event.attrs} />}
+      <ModuleContent {...event.attrs} />
+      <Module size="xs" />
     </>
   );
 }